// websites/abaatechsolutions/src/includes/types.ts
export interface APIResponseSchema {
    status: boolean;
    message: string;
    data?: any;
    code?: "SUCCESS" | "CHANGE_DEVICE" | "FAILED" | "INACTIVE" | "INVALID_ACCESS_TOKEN";
}
// ...
export interface LoginProp {
    email:string;
    password:string;
}